requirements.txt
# This own package
-e .

# Code style
black

# Testing
tox
pytest
pytest-django
pytest-pythonpath

# Package publication
setuptools
wheel
twine
